backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
spacy>=3.7.0
textblob==0.17.1
twilio==8.10.0
Pillow>=10.0.0
gunicorn==21.2.0
